Refactored registration servicing to be validation-specific
